requirements.txt
reportlab>=4.0.0
Pillow>=9.0.0
pyinstaller>=5.0.0
